+++ docs/docs_src/kafka/plaintext_security/app.py
@@ -1,8 +1,8 @@
 import ssl
 
 from faststream import FastStream
-from faststream.broker.security import SASLPlaintext
 from faststream.kafka import KafkaBroker
+from faststream.security import SASLPlaintext
 
 ssl_context = ssl.create_default_context()
 security = SASLPlaintext(ssl_context=ssl_context, username="admin", password="password")
